law-chatbot-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymongo==4.10.1
pydantic==2.10.6
python-dotenv==1.0.1

# Agent dependencies
numpy==1.26.4
google-generativeai==0.8.3
chromadb==0.5.23
qdrant-client==1.12.1
sentence-transformers==3.3.1
